+++ src/components/Packages/PackageList.js
@@ -27,8 +27,10 @@ class PackageList extends Component {
     async getPackages() {
         let packages = await getPackagesStateless();
         this.props.setDtds(packages);
+        this.props.setBiopsyIds(packages);
         this.setState({ packages: packages, unfilteredPackages: packages });
     }
+
 
     componentWillUnmount() {
         this._isMounted = false;
@@ -62,6 +64,12 @@ class PackageList extends Component {
             && this.state.packages.constructor === Array && Object.keys(this.state.packages).length === 0;
     }
 
+    setPackageLocked(index) {
+        let newPackageList = [...this.state.packages];
+        newPackageList[index].state.state = "UPLOAD_LOCKED";
+        this.setState({ packages: newPackageList });
+    }
+
     render() {
         let message = null,
             panels = [];
@@ -75,8 +83,13 @@ class PackageList extends Component {
         }
 
         else {
-            panels = this.state.packages.map((uploadPackage, index) => {
-                return <PackagePanelContainer key={index} index={index} uploadPackage={uploadPackage}/>;
+            
+            panels = applyFilters(this.props.filtering.filters, this.state.unfilteredPackages, this.props.filtering.packageTypes).map((uploadPackage, index) => {
+                return <PackagePanelContainer 
+                        key={index} 
+                        index={index} 
+                        uploadPackage={uploadPackage} 
+                        lockPackage={(index) => {this.setPackageLocked(index)}}/>;
             });
         }
 
@@ -98,6 +111,7 @@ class PackageList extends Component {
 PackageList.propTypes = {
     filtering: PropTypes.object,
     setDtds: PropTypes.func,
+    setBiopsyIds: PropTypes.func,
     poll: PropTypes.func,
     setRefresh: PropTypes.func,
     formDTD: PropTypes.object,