+++ src/components/Packages/PackagePanelContainer.js
@@ -1,5 +1,6 @@
 import { connect } from 'react-redux';
 import PackagePanel from './PackagePanel';
+import { uploadFiles, replaceFile } from "../../actions/Packages/packageActions";
 
 const mapStateToProps = (state, props) =>
 ({
@@ -8,12 +9,19 @@ const mapStateToProps = (state, props) =>
 	packageTypeIcons: state.packageTypeIcons,
 	index: props.index,
 	uploadPackage: props.uploadPackage,
+	lockPackage: props.lockPackage,
 	userInformation: state.userInformation,
 	stateDisplayMap: state.stateDisplayMap
 });
     
 const mapDispatchToProps = (dispatch, props) =>
 ({
+	uploadFiles(packageId, uploader) {
+		dispatch(uploadFiles(packageId, uploader));
+	},
+	replaceFile(packageId, fileId, uploader) {
+		dispatch(replaceFile(packageId, fileId, uploader));
+	}
 });
     
 export default connect(mapStateToProps, mapDispatchToProps)(PackagePanel);