+++ src/components/Upload/Forms/FormComponents/TextArea.js
@@ -25,7 +25,9 @@ class TextAreaComponent extends Component {
 	handleConstrainsChange = () => {
 		if (this.props.constrains !== undefined) {
 			let { resetFields } = this.props.form;
-			resetFields([this.props.constrains]);
+			this.props.constrains.forEach(constrainedField => {
+				resetFields([constrainedField]);
+			});
 		}
 	}