+++ src/components/Upload/Forms/FormComponents/TextField.js
@@ -18,7 +18,9 @@ class TextField extends Component {
 	handleConstrainsChange = () => {
 		if (this.props.constrains !== undefined) {
 			let { resetFields } = this.props.form;
-			resetFields([this.props.constrains]);
+			this.props.constrains.forEach(constrainedField => {
+				resetFields([constrainedField]);
+			});
 		}
 	}